Set up basic framework with services and views
Adds the notion of a "service" and 3 views (HomeView, LookupView, and
ManageView). New services are created by subclassing Service and setting
up the 3 Flask views (subclassed from their respective views) as
instance variables.

Each service also needs a directory in tempaltes/ with a home.html and
manage.html file.

A ServiceDirectory in the main app is used to register each service and
configure the URLs on the app. Each service is actually registered at
/service_name/* on the app - hopefully we can use mod_rewrite to map
that to different hostnames.